C++11: Mark all _CLASS_OPAQUE_REFCOUNTED classes as final.
Because _CLASS_OPAQUE_REFCOUNTED already generates a comment
telling us not to derive from them, presumably because they can
only be instantiated by reinterpret_cast<>ing a base C struct.

Ideally, _CLASS_OPAQUE_REFCOUNTED would add the final keyword,
but the class line is is not generated, so that would be a little
difficult.