Visual Studio builds: Require Visual Studio 2017
pangomm now requires C++-14 features that are only adequately supported
with Visual Studio 2017 (and later), so update the projects and related
items accordingly.  This also updates the DLL version naming to be in
line with what is in the autotools builds.

Let people know that this is the case in README.win32.