build: Add NMake Makefiles to build pangomm
This adds a set of NMake Makefiles to be used to build pangomm with
Visual Studio 2017 and later, which will share the
pango/[src|pangomm]/filelist.am with the autotools build system.  By
doing so, this will reduce the likelihood of the Visual Studio build
files getting out-of-date when source files are added or removed.

This also updates the build instructions in README.win32 (which is also
changed to DOS/Windows line endings), and will dist the NMake Makefiles
instead of the Visual Studio project files, which have been superseded
and will be removed in the next commit.