build: Add NMake Makefiles
This adds a set of NMake Makefiles, which will replace the Visual Studio
2013 projects so that the Visual Studio build files are more flexible
and easier to maintain.

Note that since this is the C++-11 version of pangomm, for Visual Studio
2015 and 2017, since they both link to the v140 Windows C/C++ runtime
DLLs, the pangomm DLL and .lib are named as pangomm-vc140-1_4.[dll|lib]
or pangomm-vc140-d-1_4.[dll|lib], since these builds can interop.