Change the ABI to pangomm-2.44
So we can use the 2.42 version numbers for stable releases in the
pangomm-1.4 ABI series.
We don't need to release a stable ABI-parallel pangomm until
we need to release gtkmm 4.0, and that won't happen until GTK+ 4.0.0
happens, and we don't know when that might be.

We've done similar ABI name changes before.