meson: Check for -mm deps manually on MSVC
The MSVC build files for the -mm library dependencies do not generate
pkg-config files, so check for them manually.  This can be eventually
replaced by the pkg-config checks when these dependencies gain
Meson build support for MSVC.

This will build and link pangomm without the import libraries, which
will require us to build and use gendef.exe that is in
MSVC_NMake/gendef, which will be done in a later commit.