meson: Relax Visual Studio check to 2013
We are able to build the C++-11 version of pangomm on Visual Studio 2013
or later, so check for that.

However, since the C++-11 support in Visual Studio is made complete in
VS 2015, display a warning message to configure the build without
warnings being fatal, as 2013 builds will spew compiler warnings.

Also, for the C++-11 versions of glibmm, cairomm and sigc++, we don't
distinguish between whether they are built by Visual Studio 2015, 2017
nor 2019, so we assume that they will all link to the same .lib's for
glibmm, cairomm and sigc++, since they should all be compatible in this
case.  In this case, Meson will warn about C++11 is not supported, but
it is safe to ignore this warning.