NMake Makefiles: Support Meson-built libsigc++
Add an option USE_MESON_LIBS for the NMake Makefile command line which
allows one to easily link to Meson-built libsigc++, as the .lib
filenames are different for libsigc++ built with NMake and Meson.

Note that when cairomm and glibmm gains Meson build support, one must
choose between Meson-built and NMake-built flavors for these as
well-mixing between flavors will not be supported, and Meson build files
will prefer Meson-built flavors of these as well.