meson.build: Look for glibmm with pkg-config on MSVC too
glibmm gained Meson build support for Visual Studio, so we can first
try to look for glibmm with pkg-config first, and then fall back to
manual searching if that cannot be found.

Note that libsigc++ also has a Meson build system as well, but either we
look for it via glibmm's pkg-config files, or we look for both glibmm
and libsigc++ manually.

Also support looking for Meson-built libsigc++ .lib's as well if glibmm
is to be looked for manually.