Meson: Allow maintainer builds on Visual Studio too
glibmm recently gained the ability to build with Visual Studio from a GIT
checkout and to install the PERL scripts used to generate the C++ sources from
the respective templates, so enable this support for pangomm as well as the
needed infrastructure are largely in place.

Note that a new configuration option has been added for Visual Studio builds
to specify the directory for gmmproc, since there is not yet Meson build
support for building glibmm, meaning glibmm built with Visual Studio does not
support generating pkg-config files.

Since some portions that we use for non-maintainer builds can be used for
maintainer builds as well, make them shared to keep additions minimal.