Meson: allow maintainer builds on Visual Studio too
glibmm recently gained the ability to build with Visual Studio from a GIT
checkout and to install the PERL scripts used to generate the C++ sources from
the respective templates, so enable this support for pangomm as well as the
needed infrastructure are largely in place.

Note that a new configuration option has been added for Visual Studio builds
to specify the directory for gmmproc, although there is Meson build support
for building glibmm, but NMake Makefiles could have been used to build
glibmm, meaning glibmm built with Visual Studio may not have pkg-config
files, so it is used to specify the location if gmmproc if necessary.

Since some portions that we use for non-maintainer builds can be used for
maintainer builds as well, make them shared to keep changes minimal.