NMake Makefiles: Detect whether we can use __declspec(dllexport)
The generated pangomm headers must be generated with a recent gmmproc so
that we may ensure that all the needed classes, functions and methods
are marked with PANGOMM_API, so that we can build pangomm without using
gendef.exe.

If the headers are generated with an older gmmproc, then we build
pangomm using gendef.exe, as we did before.