Meson/Visual Studio builds: Include toolset version by default
This makes the built DLL and .lib's contain the toolset version if the build is
carried out using Visual Studio 2017 or later, unless the
'msvc14x-parallel-installable' option is set to be false during configuration.

The reasoning behind this change is that there can be subtle problems when, for
instance, one tries to link to a Visual Studio 2017-built pangomm when building
items dependening on pangomm with Visual Studio 2019.  This is unfortunate as
Microsoft did try hard to make interoperating between binaries built with
Visual Studio 2017 and 2019 as easy as possible in terms of ABI and API, but
unfortunately this can hit the corner cases where this compatibility does not
work.

As the name suggests, this attempts to make Visual Studio 2017 and 2019 builds
share a single set of underlying C DLLs easier, while avoiding breakages caused
by such subtle differences.