NMake Makefiles: Apply toolset version for Meson-built deps
As the Meson build files for Visual Studio apply the toolset version in the
.lib filenames by default, apply the toolset version in the Meson-built -mm
.lib files that we link in, just as we did when we we link in the -mm .lib
files that was built with NMake, by default.

The option 'USE_COMPAT_LIBS' will also mean that we will use the former
behavior when we link in the Meson-built -mm .lib's, just as we did when we
link in the NMake-built -mm .lib's.