README.win32: Update Meson build info
It is now possible to build Pango directly from a GIT checkout on Visual Studio
builds for a while, so let people know about this.

Also note that glibmm and cairomm should be built with Meson in order to use the
Meson build files for pangomm, and that it is recommended to use the same compiler
to build pangomm, glibmm and cairomm.