README.win32: Convert to MarkDown
Convert to MarkDown format so that the formatting is easier on the eye, and
switch to UNIX line endings.

Also update the info in there so that the actual build situation is better
refleted, especially in regards to Visual Studio builds.  Re-organize items
for the Meson builds to be under their respective headers for MinGW and Visual
Studio.